requirements.txt
PyGithub
pyyaml
related
GitPython